WIP shuffling code around
