+++ agave_app/python/pyrenderer.cpp
@@ -108,10 +108,9 @@ OffscreenRenderer::render()
   int vw = m_fbo->width();
   int vh = m_fbo->height();
 
-  if (!m_myVolumeData.m_gesture.graphics.font.get()) {
-    m_myVolumeData.m_gesture.graphics.font.reset(new Font());
+  if (!m_myVolumeData.m_gesture.graphics.font.isLoaded()) {
     std::string fontPath = renderlib::assetPath() + "/fonts/Arial.ttf";
-    m_myVolumeData.m_gesture.graphics.font->load(fontPath.c_str());
+    m_myVolumeData.m_gesture.graphics.font.load(fontPath.c_str());
   }
 
   SceneView sceneView;
