+++ renderlib/CMakeLists.txt
@@ -26,6 +26,8 @@ target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/command.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Flags.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Flags.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Font.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Font.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Fuse.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Fuse.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/GradientData.cpp"
