+++ renderlib/graphics/CMakeLists.txt
@@ -10,6 +10,8 @@ add_subdirectory(stb)
 target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/Framebuffer.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Framebuffer.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/GestureGraphicsGL.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/GestureGraphicsGL.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/ImageXyzcGpu.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/ImageXyzcGpu.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/IRenderWindow.cpp"
