+++ renderlib/graphics/gl/CMakeLists.txt
@@ -2,8 +2,8 @@ target_include_directories(renderlib PUBLIC
 	"${CMAKE_CURRENT_SOURCE_DIR}"
 )
 target_sources(renderlib PRIVATE
-	"${CMAKE_CURRENT_SOURCE_DIR}/Font.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/Font.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/FontGL.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/FontGL.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Image2D.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Image2D.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Util.cpp"
