+++ renderlib/graphics/gl/FontGL.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <glad/glad.h> // for gl types
+
+#include "renderlib/Font.h"
+
+#include <string>
+
+class FontGL
+{
+public:
+  FontGL();
+  ~FontGL();
+
+  void load(const Font& font);
+  void unload();
+
+  GLuint getTextureID() const { return m_texID; }
+
+private:
+  GLuint m_texID;
+};
